+++ src/main/java/com/interior/application/businesss/dto/CreateBusinessServiceDto.java
@@ -1,12 +1,14 @@
 package com.interior.application.businesss.dto;
 
+import java.math.BigDecimal;
+
 public class CreateBusinessServiceDto {
 
     public record CreateBusinessMaterialDto(
         String name,
         String usageCategory,
         String category,
-        int amount,
+        BigDecimal amount,
         String unit,
         String memo,
         String materialCostPerUnit,
